refactor: Replace existing toast with yToast
