+++ packages/web-lib/components/ModalLogin.tsx
@@ -1,6 +1,6 @@
 import React, {cloneElement} from 'react';
-import {toast} from 'react-hot-toast';
 import {Modal} from '@yearn-finance/web-lib/components/Modal';
+import {yToast} from '@yearn-finance/web-lib/components/yToast';
 import {useWeb3} from '@yearn-finance/web-lib/contexts/useWeb3';
 import {useInjectedWallet} from '@yearn-finance/web-lib/hooks/useInjectedWallet';
 import IconWalletCoinbase from '@yearn-finance/web-lib/icons/IconWalletCoinbase';
@@ -18,6 +18,7 @@ function ModalLogin(props: TModalLogin): ReactElement {
 	const	{isOpen, onClose} = props;
 	const	{onConnect} = useWeb3();
 	const	detectedWalletProvider = useInjectedWallet();
+	const 	{toast} = yToast();
 
 	return (
 		<Modal
@@ -28,7 +29,7 @@ function ModalLogin(props: TModalLogin): ReactElement {
 					onClick={(): void => {
 						onConnect(
 							'INJECTED',
-							(): string => toast.error('Unsupported network. Please use Ethereum mainnet.'),
+							(): string => toast({content: 'Unsupported network. Please use Ethereum mainnet.', type: 'error'}),
 							(): void => onClose()
 						);
 					}}
@@ -40,7 +41,7 @@ function ModalLogin(props: TModalLogin): ReactElement {
 					onClick={(): void => {
 						onConnect(
 							'WALLET_CONNECT',
-							(): string => toast.error('Invalid chain'),
+							(): string => toast({content: 'Invalid chain', type: 'error'}),
 							(): void => onClose()
 						);
 					}}
@@ -52,7 +53,7 @@ function ModalLogin(props: TModalLogin): ReactElement {
 					onClick={(): void => {
 						onConnect(
 							'EMBED_COINBASE',
-							(): string => toast.error('Invalid chain'),
+							(): string => toast({content: 'Invalid chain', type: 'error'}),
 							(): void => onClose()
 						);
 					}}
@@ -64,7 +65,7 @@ function ModalLogin(props: TModalLogin): ReactElement {
 					onClick={(): void => {
 						onConnect(
 							'EMBED_GNOSIS_SAFE',
-							(): string => toast.error('Invalid chain'),
+							(): string => toast({content: 'Invalid chain', type: 'error'}),
 							(): void => onClose()
 						);
 					}}
